requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
dnspython==2.7.0
pymongo==4.10.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
